requirements.txt
fastapi==0.120.4
uvicorn==0.38.0
pillow==12.0.0
ultralytics==8.3.223
ultralytics-thop==2.0.18
pydantic-settings==2.11.0
python-multipart==0.0.20
